Move generate_complete_pairwise_dataset to util.py related to learners
Similar to the previous commits, this function is used only by learners and can therefore be moved.
